rename service to contao.util.simple_token_parser
 - move class to Contao\CoreBundle\Util\SimpleTokenParser
 - move tests to Contao\CoreBundle\Tests\Util\SimpleTokenParserTest
