+++ newsletter-bundle/src/Resources/contao/modules/ModuleSubscribe.php
@@ -376,7 +376,7 @@ class ModuleSubscribe extends Module
 
 		// Send the token
 		$text = System::getContainer()
-			->get('contao.simple_tokens.parser')
+			->get('contao.util.simple_token_parser')
 			->parseTokens($this->nl_subscribe, $arrData);
 
 		$optInToken->send(sprintf($GLOBALS['TL_LANG']['MSC']['nl_subject'], Idna::decode(Environment::get('host'))), $text);
